separate loader from inspector to hold SRP
