+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/UpdaterServiceDelegatee.scala
@@ -5,7 +5,7 @@ import android.content.Intent
 import android.os.IBinder
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.glue.service.ServiceControl
-import x7c1.linen.repository.crawler.SourceUpdaterQueue
+import x7c1.linen.repository.crawler.{RemoteSourceLoader, SourceUpdaterQueue}
 import x7c1.linen.scene.updater.UpdaterMethods
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.service.CommandStartType
@@ -14,7 +14,8 @@ import x7c1.wheat.modern.decorator.service.CommandStartType.NotSticky
 
 class UpdaterServiceDelegatee(service: Service with ServiceControl){
   private lazy val helper = new DatabaseHelper(service)
-  private lazy val queue = new SourceUpdaterQueue(helper)
+
+  private lazy val queue = new SourceUpdaterQueue(helper, RemoteSourceLoader)
 
   def onBind(intent: Intent): Option[IBinder] = {
     Log info "[init]"
