+++ linen-repository/src/main/scala/x7c1/linen/repository/crawler/SourceInspector.scala
@@ -1,26 +1,15 @@
 package x7c1.linen.repository.crawler
 
-import java.io.{BufferedInputStream, InputStreamReader}
-import java.net.{HttpURLConnection, URL}
+import java.net.URL
 
-import com.google.code.rome.android.repackaged.com.sun.syndication.feed.synd.{SyndEntry, SyndFeed}
-import com.google.code.rome.android.repackaged.com.sun.syndication.io.SyndFeedInput
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.struct.{EntryParts, SourceRecord}
-import x7c1.linen.repository.date.Date
-import x7c1.linen.repository.entry.EntryUrl
-import x7c1.wheat.modern.callback.CallbackTask
-import x7c1.wheat.modern.callback.CallbackTask.task
-import x7c1.wheat.modern.callback.TaskProvider.using
-
-import scala.concurrent.Future
+import x7c1.linen.database.struct.SourceRecord
 
 object SourceInspector {
   def apply(helper: DatabaseHelper ): SourceInspector = new SourceInspector(helper)
 }
 
 class SourceInspector private (helper: DatabaseHelper){
-  import Implicits._
 
   def inspectSource(sourceId: Long): Either[SourceInspectorError, InspectedSource] =
     helper.readable.find[SourceRecord].by(sourceId) via {
@@ -28,50 +17,6 @@ class SourceInspector private (helper: DatabaseHelper){
       case Right(None) => Left(SourceNotFound(sourceId))
       case Right(Some(source)) => Right(InspectedSource(source))
     }
-
-  def loadSource(source: InspectedSource): Future[LoadedSource] = {
-    import collection.JavaConverters._
-
-    Future(source.feedUrl).map(loadRawFeed).flatMap(_.toFuture) map { feed =>
-      val entries = feed.getEntries.asScala map { case x: SyndEntry => x }
-      new LoadedSource(
-        sourceId = source.sourceId,
-        title = Option(feed.getTitle) getOrElse "",
-        description = Option(feed.getDescription) getOrElse "",
-        entries = entries map convertEntry(source.sourceId)
-      )
-    }
-  }
-  private def loadRawFeed(feedUrl: URL): CallbackTask[SyndFeed] = {
-    for {
-      connection <- task {
-        val connection = feedUrl.openConnection().asInstanceOf[HttpURLConnection]
-        connection setRequestMethod "GET"
-        connection
-      }
-      stream <- using(new BufferedInputStream(connection.getInputStream))
-      reader <- using(new InputStreamReader(stream))
-    } yield {
-      new SyndFeedInput().build(reader)
-    }
-  }
-
-  private def convertEntry(sourceId: Long)(entry: SyndEntry): Either[InvalidEntry, EntryParts] = {
-    try for {
-      url <- (Option(entry.getLink) toRight EmptyUrl()).right
-      published <- (Option(entry.getPublishedDate) toRight EmptyPublishedDate()).right
-    } yield EntryParts(
-      sourceId = sourceId,
-      title = Option(entry.getTitle) getOrElse "",
-      content = Option(entry.getDescription.getValue) getOrElse "",
-      author = Option(entry.getAuthor) getOrElse "",
-      url = EntryUrl(url),
-      createdAt = Date(published)
-    ) catch {
-      case e: Exception => Left(Abort(e))
-    }
-  }
-
 }
 
 case class InspectedSource(
